+++ src/RayTracer/VCM.cpp
@@ -118,7 +118,7 @@ void VCM::init() {
 }
 
 void VCM::render() {
-	const float radius_factor = json_util::get_or(integrator_config, "radius_factor", 1.);//static_cast<float>(integrator_config["radius_factor"]);
+	const float radius_factor = static_cast<float>(json_util::get_or(integrator_config, "radius_factor", 1.));//static_cast<float>(integrator_config["radius_factor"]);
 	const int enable_vm 	  = json_util::get_or(integrator_config, "enable_vm", 1);
 
 	CommandBuffer cmd(&instance->vkb.ctx, /*start*/ true, VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT);
